Refactor speech interruption.
